chore: deprecate InputTextField
