+++ frontend/app_flowy/lib/plugins/grid/presentation/widgets/cell/checklist_cell/checklist_prograss_bar.dart
@@ -56,6 +56,7 @@ class _SliverChecklistPrograssBarDelegate
             children: [
               FlowyTextField(
                 autoClearWhenDone: true,
+                submitOnLeave: true,
                 hintText: LocaleKeys.grid_checklist_panelTitle.tr(),
                 onChanged: (text) {
                   context
